convert cache_ops + cuda_utils

+++ csrc/cuda_utils.h
@@ -2,6 +2,6 @@
 
 #include <torch/extension.h>
 
-int get_device_attribute(int attribute, int device_id);
+int64_t get_device_attribute(int64_t attribute, int64_t device_id);
 
-int get_max_shared_memory_per_block_device_attribute(int device_id);
+int64_t get_max_shared_memory_per_block_device_attribute(int64_t device_id);
